SearchDataTransformer: centralize allResultsForVertical parsing
This commit moves _formResponseFromAllResultsForVertical from models/verticalresults.js
up one level into searchdatatransformer, so that other models for the same set of request,
like DYNAMIC_FILTERS, do not have to redo this parsing.

The only logical change of this commit is that before verticalresults determined
it was using allResultsForVertical if response.results && response.results.length > 0;
was false, and dynamicfilters used allResultsForVertical if
response.resultsCount === 0 && response.allResultsForVertical was true. Now they both
use the former.

TEST=manual
checked that can run searches and apply facets for both normal and no results.
