+++ tests/core/search/searchdatatransformer.js
@@ -7,6 +7,7 @@ import DynamicFilters from '../../../src/core/models/dynamicfilters';
 import SearchIntents from '../../../src/core/models/searchintents';
 import LocationBias from '../../../src/core/models/locationbias';
 import Response from '../../fixtures/responseWithResults.json';
+import ResultsContext from '../../../src/core/storage/resultscontext';
 
 describe('tranform vertical search response', () => {
   let response;
@@ -18,17 +19,129 @@ describe('tranform vertical search response', () => {
   it('transforms vertical response correctly', () => {
     const data = response;
     const formatters = [];
-    const result = SearchDataTransformer.transformVertical(data, formatters);
+    const result = SearchDataTransformer.transformVertical(data, formatters, 'vkey');
     expect(result).toMatchObject(
       {
         [StorageKeys.QUERY_ID]: data.response.queryId,
         [StorageKeys.NAVIGATION]: new Navigation(), // Vertical doesn't respond with ordering, so use empty nav.
-        [StorageKeys.VERTICAL_RESULTS]: VerticalResults.from(data.response, formatters),
-        [StorageKeys.DYNAMIC_FILTERS]: DynamicFilters.from(data.response),
+        [StorageKeys.VERTICAL_RESULTS]: VerticalResults.from(data.response, formatters, 'vkey', ResultsContext.NORMAL),
+        [StorageKeys.DYNAMIC_FILTERS]: DynamicFilters.from(data.response.facets),
         [StorageKeys.INTENTS]: SearchIntents.from(data.response.searchIntents),
         [StorageKeys.SPELL_CHECK]: SpellCheck.from(data.response.spellCheck),
         [StorageKeys.LOCATION_BIAS]: LocationBias.from(data.response.locationBias)
       }
     );
+  });
+});
+
+describe('forming no results response', () => {
+  let response;
+
+  beforeEach(() => {
+    response = {
+      results: [],
+      resultsCount: 0,
+      facets: [
+        {
+          'fieldId': 'c_features',
+          'displayName': 'Features',
+          'options': []
+        }
+      ],
+      allResultsForVertical: {
+        resultsCount: 2,
+        results: [
+          {
+            data: {
+              id: 'faq-0',
+              type: 'faq',
+              answer: 'black',
+              name: 'what is alexis\' favorite color?'
+            },
+            highlightedFields: {}
+          }
+        ],
+        facets: [
+          {
+            'fieldId': 'c_features',
+            'displayName': 'Features',
+            'options': [
+              {
+                'displayName': 'Dog Friendly',
+                'count': 1,
+                'selected': false,
+                'filter': {
+                  'c_features': {
+                    '$eq': 'Dog Friendly'
+                  }
+                }
+              }
+            ]
+          }
+        ]
+      },
+      appliedQueryFilters: []
+    };
+  });
+
+  it('does not alter original response ', () => {
+    const initialResponse = { ...response };
+    SearchDataTransformer._formResponseWithContext(response, ResultsContext.NO_RESULTS);
+
+    expect(response).toEqual(initialResponse);
+  });
+
+  it('properly converts response with data', () => {
+    const convertedResponse = SearchDataTransformer._formResponseWithContext(response, ResultsContext.NO_RESULTS);
+
+    expect(convertedResponse).toEqual({
+      resultsCount: 2,
+      facets: response.allResultsForVertical.facets,
+      results: [
+        {
+          data: {
+            id: 'faq-0',
+            type: 'faq',
+            answer: 'black',
+            name: 'what is alexis\' favorite color?'
+          },
+          highlightedFields: {}
+        }
+      ],
+      allResultsForVertical: response.allResultsForVertical,
+      appliedQueryFilters: []
+    });
+  });
+
+  it('properly converts response with empty results', () => {
+    const responseEmptyResults = {
+      results: [],
+      resultsCount: 0,
+      allResultsForVertical: [],
+      appliedQueryFilters: []
+    };
+    const convertedResponse = SearchDataTransformer._formResponseWithContext(responseEmptyResults, ResultsContext.NO_RESULTS);
+
+    expect(convertedResponse).toEqual({
+      results: [],
+      resultsCount: 0,
+      allResultsForVertical: [],
+      appliedQueryFilters: []
+    });
+  });
+
+  it('properly converts response with non-present results', () => {
+    const responseEmptyResults = {
+      results: [],
+      resultsCount: 0,
+      appliedQueryFilters: []
+    };
+    const convertedResponse = SearchDataTransformer._formResponseWithContext(responseEmptyResults, ResultsContext.NO_RESULTS);
+
+    expect(convertedResponse).toEqual({
+      results: [],
+      resultsCount: 0,
+      appliedQueryFilters: []
+    });
   });
 });
